assignment/Assignment16/requirements.txt
streamlit
requests
beautifulsoup4
sentence-transformers
faiss-cpu
transformers
torch
numpy